[host_callback] Change host_callback lowering to use custom calls on GPU.
Currently, host_callback uses outfeed lowering for TPU and GPU and custom_call for CPU (by way of mlir.emit_python_callback). We want to align it with jax.debug APIs, which use custom_call on GPU also. This change should not affect TPU behavior, we'll change that in a subsequent change.

You can still use the `--jax_host_callback_use_outfeed=True` flag to get the previous behavior.

This is part of a longer project to replace uses of host_callback with jax.pure_callback and jax.experimental.io_callback.

PiperOrigin-RevId: 557405952